+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Add a `microfloat` feature for expanded subfloat and complex subfloat data type / element support
+  - The `float8` feature is considered deprecated as the associated types are not IEEE 754-compliant
 
 ## [0.23.10](https://github.com/zarrs/zarrs/releases/tag/zarrs-v0.23.10) - 2026-04-09