+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add a `microfloat` feature for expanded subfloat and complex subfloat data type / element support
   - The `float8` feature is considered deprecated as the associated types are not IEEE 754-compliant
 
+### Changed
+- Bump `zarrs_metadata_ext` to 0.4.3
+
 ## [0.23.10](https://github.com/zarrs/zarrs/releases/tag/zarrs-v0.23.10) - 2026-04-09
 
 ### Added